Backend/requirements.txt
bidict==0.22.1
asttokens==2.4.1
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
dnspython==2.5.0
executing==2.0.1
Flask==3.0.1
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
google-api-core==2.16.2
google-auth==2.27.0
google-cloud-aiplatform==1.40.0
google-cloud-bigquery==3.17.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
numpy==1.26.3
packaging==23.2
parso==0.8.3
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pymongo==4.6.1
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.8.2
python-socketio==5.11.0
simple-websocket==1.0.0
requests==2.31.0
rsa==4.9
shapely==2.0.2
six==1.16.0
stack-data==0.6.3
traitlets==5.14.1
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
wsproto==1.2.0
google-genai==1.59.0